apps/api/pyproject.toml
[project]
name = "zeee-api"
version = "0.1.0"
description = "Zeee Pitch Zooo API backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn",
    "python-dotenv",
    "pydantic-settings",
    "supabase",
    "llama-index-core>=0.14.18",
    "llama-index-readers-file>=0.6.0",
    "llama-index-embeddings-openai>=0.6.0",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.22",
    "pymupdf>=1.27.2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
]
